requirements.txt
matplotlib==3.5.1
numpy==1.22.1
opencv_python==4.5.5.62
Pillow==9.0.0
tensorflow==2.7.0
tensorflow_macos==2.7.0
